Removed hack for key search popup
