+++ octopod-frontend/octopod-frontend.cabal
@@ -134,5 +134,6 @@ executable frontend
                      , fuzzyfind
                      , deepseq
                      , patch
+                     , reflex-dom-core
   hs-source-dirs:      src
   default-language:    Haskell2010
